move token validation to services/local (#67266)
