+++ lib/auth/init_test.go
@@ -1922,6 +1922,18 @@ spec:
   github:
     allow:
       - repository: gravitational/example`
+	badTokenYAML = `kind: token
+version: v2
+metadata:
+  name: iam-token-without-orgid-but-ous
+  expires: "3000-01-01T00:00:00Z"
+spec:
+  roles: [Node]
+  join_method: iam
+  allow:
+   - aws_account: "123456789012"
+     aws_organizational_units:
+       include: [r-rootid]`
 	roleYAML = `kind: role
 version: v7
 metadata:
@@ -1988,6 +2000,7 @@ func TestInit_ApplyOnStartup(t *testing.T) {
 
 	user := resourceFromYAML(t, userYAML).(types.User)
 	token := resourceFromYAML(t, tokenYAML).(types.ProvisionToken)
+	badToken := resourceFromYAML(t, badTokenYAML).(types.ProvisionToken)
 	role := resourceFromYAML(t, roleYAML).(types.Role)
 	lock := resourceFromYAML(t, lockYAML).(types.Lock)
 	clusterNetworkingConfig := resourceFromYAML(t, clusterNetworkingConfYAML).(types.ClusterNetworkingConfig)
@@ -2007,6 +2020,13 @@ func TestInit_ApplyOnStartup(t *testing.T) {
 				cfg.ApplyOnStartupResources = append(cfg.ApplyOnStartupResources, user)
 				cfg.ApplyOnStartupResources = append(cfg.ApplyOnStartupResources, role)
 				cfg.ApplyOnStartupResources = append(cfg.ApplyOnStartupResources, token)
+			},
+			assertError: require.Error,
+		},
+		{
+			name: "Apply invalid provision token",
+			modifyConfig: func(cfg *auth.InitConfig) {
+				cfg.ApplyOnStartupResources = append(cfg.ApplyOnStartupResources, badToken)
 			},
 			assertError: require.Error,
 		},
