+++ lib/join/oraclejoin/join_test.go
@@ -184,7 +184,7 @@ func TestJoinOracle(t *testing.T) {
 						makeCompartmentID("othercompartment"),
 						makeCompartmentID("mycompartment"),
 					},
-					Regions: []string{"otherregion", "phx"},
+					Regions: []string{"us-phoenix-1", "phx"},
 				},
 			},
 			tokenName:        "mytoken",
@@ -205,7 +205,7 @@ func TestJoinOracle(t *testing.T) {
 						makeCompartmentID("othercompartment"),
 						makeCompartmentID("mycompartment"),
 					},
-					Regions: []string{"otherregion", "phx"},
+					Regions: []string{"us-phoenix-1", "phx"},
 					Instances: []string{
 						makeInstanceID("phx", "otherinstance"),
 						makeInstanceID("phx", "myinstance"),
@@ -304,7 +304,7 @@ func TestJoinOracle(t *testing.T) {
 						makeCompartmentID("othercompartment"),
 						makeCompartmentID("mycompartment"),
 					},
-					Regions: []string{"otherregion", "phx"},
+					Regions: []string{"us-phoenix-1", "phx"},
 				},
 			},
 			tokenName:        "mytoken",
@@ -325,7 +325,7 @@ func TestJoinOracle(t *testing.T) {
 						makeCompartmentID("othercompartment"),
 						makeCompartmentID("mycompartment"),
 					},
-					Regions: []string{"otherregion", "phx"},
+					Regions: []string{"us-phoenix-1", "phx"},
 					Instances: []string{
 						makeInstanceID("phx", "otherinstance"),
 						makeInstanceID("phx", "myinstance"),
