+++ lib/services/local/export_test.go
@@ -18,7 +18,11 @@
 
 package local
 
-import "time"
+import (
+	"time"
+
+	"github.com/gravitational/teleport/api/types"
+)
 
 // SessionDataLimiter exports sdLimiter for tests.
 var SessionDataLimiter = sdLimiter
@@ -37,3 +41,11 @@ const (
 	UsersPrefix  = usersPrefix
 	ParamsPrefix = paramsPrefix
 )
+
+func ValidateOracleJoinToken(token types.ProvisionToken) error {
+	return validateOracleJoinToken(token)
+}
+
+func ValidateProvisionToken(token types.ProvisionToken) error {
+	return validateProvisionToken(token)
+}
